Destroy session in DefaultSmppClient on bind error
In case bind fails the caller wont be able to destroy the created
session (if any). In the case of DefaultSmppSession this will
cause the WindowMonitor of the sendWindow to detect a missing
Window.destroy and log an error about this.

This patch destroys the session in cases the session has been
created but wont be returned to the caller.