fixed tests to have a session field which gets closed in @After, so that there's no need to have a try/finally block in each test. added tests for ensuring that a client can bind to a specific local address (host and maybe also port).
For the test that checks that a client is bound to a specific port as well, a port is chosen at random.
However, it is again checked that if a BindException occurs due to trying to bind to a port that was already in use earlier, the test is ignored (with junit's Assume), rather than failing, since there's not much to do for tcp connections in TIME_WAIT state.